Make preprocessing cmssw-independent.
