+++ .gitignore
@@ -14,3 +14,4 @@ __init__.py
 /soft
 .vscode
 /.law
+/data
