+++ LawWorkflows/law.cfg
@@ -1,9 +1,9 @@
 [modules]
 
-law.ShuffleMergeSpectral
-law.Hadd
-law.Training
-law.FeatureScaling
+LawWorkflows.ShuffleMergeSpectral
+LawWorkflows.Hadd
+LawWorkflows.Training
+LawWorkflows.FeatureScaling
 
 
 [job]
